+++ package-release.ps1
@@ -6,7 +6,7 @@ $here = Split-Path -Parent $MyInvocation.MyCommand.Path
 $releaseDir = Join-Path $here 'release'
 New-Item -ItemType Directory -Force -Path $releaseDir | Out-Null
 
-$version = '3.2.0'
+$version = '3.2.1'
 Copy-Item (Join-Path $here 'EventMsg.exe') (Join-Path $releaseDir 'EventMsg.exe') -Force
 Copy-Item (Join-Path $here 'EventMsg-Dev.exe') (Join-Path $releaseDir 'EventMsg-Dev.exe') -Force
 Write-Host "Release v$version ->"